Split gen_cpp into three files
The single file was getting way to long...
