+++ src/generate/gen_cpp.cpp
@@ -32,36 +32,6 @@
 
 using namespace code;
 
-extern const char* cpp_rust_end_cmt_line;  // "// ************* End of generated code"
-
-const std::map<wxBitmapType, std::string_view> g_map_handlers = {
-    { wxBITMAP_TYPE_ICO, "wxICOHandler" },   { wxBITMAP_TYPE_CUR, "wxCURHandler" },
-    { wxBITMAP_TYPE_XPM, "wxXPMHandler" },
-#ifndef __WXOSX__
-    { wxBITMAP_TYPE_TIFF, "wxTIFFHandler" },
-#endif
-    { wxBITMAP_TYPE_GIF, "wxGIFHandler" },   { wxBITMAP_TYPE_PNG, "wxPNGHandler" },
-    { wxBITMAP_TYPE_JPEG, "wxJPEGHandler" }, { wxBITMAP_TYPE_PNM, "wxPNMHandler" },
-    { wxBITMAP_TYPE_ANI, "wxANIHandler" },   { wxBITMAP_TYPE_WEBP, "wxWEBPHandler" }
-};
-
-const std::map<wxBitmapType, std::string_view> g_map_types = {
-    { wxBITMAP_TYPE_BMP, "wxBITMAP_TYPE_BMP" },
-    { wxBITMAP_TYPE_ICO, "wxBITMAP_TYPE_ICO" },
-    { wxBITMAP_TYPE_CUR, "wxBITMAP_TYPE_CUR" },
-    { wxBITMAP_TYPE_XPM, "wxBITMAP_TYPE_XPM" },
-    { wxBITMAP_TYPE_ICO_RESOURCE, "wxBITMAP_TYPE_ICO_RESOURCE" },
-#ifndef __WXOSX__
-    { wxBITMAP_TYPE_TIFF, "wxBITMAP_TYPE_TIFF" },
-#endif
-    { wxBITMAP_TYPE_GIF, "wxBITMAP_TYPE_GIF" },
-    { wxBITMAP_TYPE_PNG, "wxBITMAP_TYPE_PNG" },
-    { wxBITMAP_TYPE_JPEG, "wxBITMAP_TYPE_JPEG" },
-    { wxBITMAP_TYPE_PNM, "wxBITMAP_TYPE_PNM" },
-    { wxBITMAP_TYPE_ANI, "wxBITMAP_TYPE_ANI" },
-    { wxBITMAP_TYPE_WEBP, "wxBITMAP_TYPE_WEBP" }
-};
-
 namespace
 {
     class GenData
@@ -102,6 +72,36 @@ namespace
 {
     void GenCppForm(GenData& gen_data, Node* form);
 }
+
+extern const char* cpp_rust_end_cmt_line;  // "// ************* End of generated code"
+
+const std::map<wxBitmapType, std::string_view> g_map_handlers = {
+    { wxBITMAP_TYPE_ICO, "wxICOHandler" },   { wxBITMAP_TYPE_CUR, "wxCURHandler" },
+    { wxBITMAP_TYPE_XPM, "wxXPMHandler" },
+#ifndef __WXOSX__
+    { wxBITMAP_TYPE_TIFF, "wxTIFFHandler" },
+#endif
+    { wxBITMAP_TYPE_GIF, "wxGIFHandler" },   { wxBITMAP_TYPE_PNG, "wxPNGHandler" },
+    { wxBITMAP_TYPE_JPEG, "wxJPEGHandler" }, { wxBITMAP_TYPE_PNM, "wxPNMHandler" },
+    { wxBITMAP_TYPE_ANI, "wxANIHandler" },   { wxBITMAP_TYPE_WEBP, "wxWEBPHandler" }
+};
+
+const std::map<wxBitmapType, std::string_view> g_map_types = {
+    { wxBITMAP_TYPE_BMP, "wxBITMAP_TYPE_BMP" },
+    { wxBITMAP_TYPE_ICO, "wxBITMAP_TYPE_ICO" },
+    { wxBITMAP_TYPE_CUR, "wxBITMAP_TYPE_CUR" },
+    { wxBITMAP_TYPE_XPM, "wxBITMAP_TYPE_XPM" },
+    { wxBITMAP_TYPE_ICO_RESOURCE, "wxBITMAP_TYPE_ICO_RESOURCE" },
+#ifndef __WXOSX__
+    { wxBITMAP_TYPE_TIFF, "wxBITMAP_TYPE_TIFF" },
+#endif
+    { wxBITMAP_TYPE_GIF, "wxBITMAP_TYPE_GIF" },
+    { wxBITMAP_TYPE_PNG, "wxBITMAP_TYPE_PNG" },
+    { wxBITMAP_TYPE_JPEG, "wxBITMAP_TYPE_JPEG" },
+    { wxBITMAP_TYPE_PNM, "wxBITMAP_TYPE_PNM" },
+    { wxBITMAP_TYPE_ANI, "wxBITMAP_TYPE_ANI" },
+    { wxBITMAP_TYPE_WEBP, "wxBITMAP_TYPE_WEBP" }
+};
 
 // clang-format off
 
@@ -423,44 +423,6 @@ bool GenerateCppFiles(GenResults& results, std::vector<tt_string>* pClassList)
     return results.updated_files.size() > 0;
 }
 
-void CppCodeGenerator::GenHdrNameSpace(tt_string& namespace_prop, tt_string_vector& names,
-                                       size_t& indent)
-{
-    // namespace_prop can be a single or multiple namespaces separated by either :: or ;. Replace
-    // both separator types with a single ':' character.
-    namespace_prop.Replace("::", ":");
-    namespace_prop.Replace(";", ":");
-    names.SetString(namespace_prop, ':');
-
-    tt_string using_name;
-    m_header->writeLine();  // start with a blank line
-    for (auto& iter: names)
-    {
-        m_header->writeLine(tt_string() << "namespace " << iter);
-        m_header->writeLine("{");
-        m_header->Indent();
-        // This lets the caller know how much to indent the code inside the namespace
-        ++indent;
-
-        if (using_name.empty())
-        {
-            using_name = "using namespace ";
-        }
-        else
-        {
-            using_name += "::";
-        }
-        using_name += iter;
-    }
-    m_header->SetLastLineBlank();
-
-    if (using_name.size())
-    {
-        using_name << ';';
-        m_source->writeLine(using_name);
-    }
-}
-
 void CppCodeGenerator::GenCppImageFunctions()
 {
     // First, generate the header files needed
@@ -553,141 +515,6 @@ void CppCodeGenerator::GenCppImageFunctions()
     {
         Code code(m_form_node, GEN_LANG_CPLUSPLUS);
         WriteImageConstruction(code);
-    }
-}
-
-void CppCodeGenerator::GenInitHeaderFile(std::set<std::string>& hdr_includes)
-{
-    std::vector<std::string> ordered_includes;
-    if (auto pos = hdr_includes.find("#include <wx/generic/stattextg.h>");
-        pos != hdr_includes.end())
-    {
-        hdr_includes.erase(pos);
-        if (pos = hdr_includes.find("#include <wx/stattext.h>"); pos != hdr_includes.end())
-        {
-            hdr_includes.erase(pos);
-        }
-
-        if (ordered_includes.empty())
-        {
-            ordered_includes.emplace_back("// Order dependent includes");
-        }
-
-        ordered_includes.emplace_back("#include <wx/stattext.h>");
-        ordered_includes.emplace_back("#include <wx/generic/stattextg.h>");
-    }
-    if (auto pos = hdr_includes.find("#include <wx/generic/treectlg.h>"); pos != hdr_includes.end())
-    {
-        hdr_includes.erase(pos);
-        if (pos = hdr_includes.find("#include <wx/treectrl.h>"); pos != hdr_includes.end())
-        {
-            hdr_includes.erase(pos);
-        }
-
-        if (ordered_includes.empty())
-        {
-            ordered_includes.emplace_back("// Order dependent includes");
-        }
-
-        ordered_includes.emplace_back("#include <wx/treectrl.h>");
-        ordered_includes.emplace_back("#include <wx/generic/treectlg.h>");
-    }
-
-    if (ordered_includes.size())
-    {
-        for (auto& iter: ordered_includes)
-        {
-            m_header->writeLine(iter);
-        }
-        m_header->writeLine();
-    }
-
-    // First output all the wxWidget header files
-    for (const auto& iter: hdr_includes)
-    {
-        if (tt::contains(iter, "<wx"))
-        {
-            m_header->writeLine(iter);
-        }
-    }
-
-    m_header->writeLine();
-
-    std::vector<std::string> namespaces;
-    for (auto iter = hdr_includes.begin(); iter != hdr_includes.end();)
-    {
-        if (iter->starts_with("namespace "))
-        {
-            namespaces.emplace_back(*iter);
-            iter = hdr_includes.erase(iter);
-        }
-        else
-        {
-            ++iter;
-        }
-    }
-
-    // Now output all the other header files (this will include forward class declarations)
-    for (const auto& iter: hdr_includes)
-    {
-        if (!tt::contains(iter, "<wx"))
-        {
-            m_header->writeLine(iter);
-        }
-    }
-
-    m_header->writeLine();
-
-    if (m_form_node->HasValue(prop_header_preamble))
-    {
-        WritePropHdrCode(m_form_node, prop_header_preamble);
-    }
-
-    if (m_form_node->HasValue(prop_system_hdr_includes))
-    {
-        m_header->writeLine();
-        tt_view_vector list;
-        list.SetString(m_form_node->as_string(prop_system_hdr_includes));
-        for (auto& iter: list)
-        {
-            m_header->writeLine(tt_string("#include <") << iter << '>');
-        }
-    }
-
-    if (m_form_node->HasValue(prop_local_hdr_includes))
-    {
-        m_header->writeLine();
-        tt_view_vector list;
-        list.SetString(m_form_node->as_string(prop_local_hdr_includes));
-        for (auto& iter: list)
-        {
-            m_header->writeLine(tt_string("#include \"") << iter << '"');
-        }
-    }
-
-    if (namespaces.size())
-    {
-        m_header->writeLine();
-        for (auto& iter: namespaces)
-        {
-            tt_view_vector list(iter, '\n');
-
-            // See gen_custom_ctrl.cpp -- GetIncludes(). Format is namespace name\n{\nclass name;\n}
-            m_header->writeLine(list[0]);
-            m_header->writeLine(list[1]);
-            m_header->Indent();
-
-            for (size_t idx = 2; idx < list.size(); ++idx)
-            {
-                if (list[idx].starts_with("}"))
-                {
-                    m_header->Unindent();
-                    m_header->writeLine(list[idx]);
-                    break;
-                }
-                m_header->writeLine(list[idx]);
-            }
-        }
     }
 }
 
@@ -871,313 +698,6 @@ void CppCodeGenerator::GenerateClass(PANEL_PAGE panel_type)
     {
         code.Eol().Str("#endif  // ").Str(m_form_node->as_string(prop_cpp_conditional));
         m_source->writeLine(code);
-    }
-}
-
-void CppCodeGenerator::GenerateCppClassHeader(bool class_namespace)
-{
-    ASSERT(m_language == GEN_LANG_CPLUSPLUS);
-
-    if (m_form_node->is_Gen(gen_Images) || m_form_node->is_Gen(gen_Data))
-    {
-        // There is a header for this, but it's not a class header
-        return;
-    }
-
-    if (!m_form_node->HasValue(prop_class_name))
-    {
-        FAIL_MSG(tt_string("Missing \"name\" property in ") << m_form_node->get_DeclName());
-        return;
-    }
-
-    auto* generator = m_form_node->get_NodeDeclaration()->get_Generator();
-    Code code(m_form_node, GEN_LANG_CPLUSPLUS);
-
-    // This may result in two blank lines, but without it there may be a case where there is no
-    // blank line at all.
-    m_header->writeLine();
-
-    if (generator->PreClassHeaderCode(code))
-    {
-        m_header->writeLine(code);
-        code.clear();
-    }
-
-    // If the class has a namespace, then this was already written
-    if (!class_namespace && m_embedded_images.size())
-    {
-        WriteImagePostHeader();
-        m_header->writeLine();
-    }
-
-    code.Str("class ");
-    if (m_form_node->HasValue(prop_class_decoration))
-    {
-        code.as_string(prop_class_decoration) += " ";
-    }
-    code.as_string(prop_class_name) += " : public ";
-    if (generator->BaseClassNameCode(code))
-    {
-        if (m_form_node->HasValue(prop_additional_inheritance))
-        {
-            tt_string_vector class_list(m_form_node->as_string(prop_additional_inheritance), '"',
-                                        tt::TRIM::both);
-            for (auto& iter: class_list)
-            {
-                code.Str(", public ").Str(iter);
-            }
-        }
-        m_header->writeLine(code);
-        code.clear();
-    }
-    else
-    {
-        if (!m_form_node->is_Type(type_DocViewApp) && !m_form_node->is_Type(type_wx_document) &&
-            !m_form_node->is_Type(type_wx_view))
-        {
-            FAIL_MSG("All form generators need to support BaseClassNameCode() to provide the class "
-                     "name to derive from.");
-        }
-        else
-        {
-            // The only way this would be valid is if the base class didn't derive from anything.
-            m_header->writeLine(tt_string() << "class " << m_form_node->as_string(prop_class_name));
-        }
-    }
-
-    m_header->writeLine("{");
-    m_header->writeLine("public:");
-    m_header->Indent();
-    m_header->SetLastLineBlank();
-
-    // The set is used to prevent duplicates and to write the lines sorted. Call WriteSetLines()
-    // to write the lines and clear the set.
-    std::set<std::string> code_lines;
-    CollectMemberVariables(m_form_node, Permission::Public, code_lines);
-    if (code_lines.size())
-    {
-        WriteSetLines(m_header, code_lines);
-        m_header->writeLine();
-    }
-
-    for (auto& member: m_map_public_members)
-    {
-        code.clear();
-        BeginPlatformCode(code, member.first);
-        m_header->writeLine(code);
-        for (const auto& member_code: member.second)
-        {
-            m_header->writeLine(member_code);
-        }
-        m_header->writeLine("#endif  // limited to specific platforms");
-    }
-
-    GenCppEnumIds(m_form_node);
-
-    if (m_form_node->as_bool(prop_const_values))
-    {
-        code.clear();
-        if (m_form_node->HasProp(prop_id))
-        {
-            code.Eol(eol_if_needed).Str("static const int form_id = ");
-            if (m_form_node->as_string(prop_id).size())
-            {
-                code.as_string(prop_id) += ";";
-            }
-            else
-            {
-                code.Str("wxID_ANY;");
-            }
-        }
-        if (m_form_node->HasProp(prop_style))
-        {
-            code.Eol(eol_if_needed).Str("static const int form_style = ");
-            if (m_form_node->as_string(prop_style).size())
-            {
-                code.as_string(prop_style) += ";";
-            }
-            else
-            {
-                code.Str("0;");
-            }
-        }
-        else if (m_form_node->HasProp(prop_window_style))
-        {
-            code.Eol(eol_if_needed).Str("static const int form_style = ");
-            if (m_form_node->as_string(prop_window_style).size())
-            {
-                code.as_string(prop_window_style) += ";";
-            }
-            else
-            {
-                code.Str("0;");
-            }
-        }
-        if (m_form_node->HasProp(prop_pos))
-        {
-            code.Eol(eol_if_needed)
-                .Str("static const wxPoint form_pos() { return ")
-                .Pos(prop_pos, no_dpi_scaling) += "; }";
-        }
-        if (m_form_node->HasProp(prop_size))
-        {
-            code.Eol(eol_if_needed)
-                .Str("static const wxSize form_size() { return  ")
-                .WxSize(prop_size, no_dpi_scaling) += "; }";
-        }
-        if (m_form_node->HasProp(prop_title))
-        {
-            code.Eol(eol_if_needed).Str("static const wxString form_title() { return ");
-            if (m_form_node->HasValue(prop_title))
-            {
-                code.Str("wxString::FromUTF8(\"").as_string(prop_title) += "\"); }";
-            }
-            else
-            {
-                code.Str("wxEmptyString; }");
-            }
-        }
-
-        if (code.size())
-        {
-            m_header->writeLine(code);
-            m_header->writeLine();
-            m_header->writeLine();
-        }
-    }
-
-    code.clear();
-    if (generator->HeaderCode(code))
-    {
-        m_header->writeLine(code);
-    }
-
-    m_header->SetLastLineBlank();
-
-    GenCppValidatorFunctions(m_form_node);
-    m_header->writeLine();
-
-    if (m_form_node->HasValue(prop_inserted_hdr_code))
-    {
-        WritePropHdrCode(m_form_node, prop_inserted_hdr_code);
-    }
-
-    if (m_form_node->HasValue(prop_class_methods))
-    {
-        tt_string_vector class_list(m_form_node->as_string(prop_class_methods), '"',
-                                    tt::TRIM::both);
-        if (class_list.size())
-        {
-            m_header->writeLine();
-            for (auto& iter: class_list)
-            {
-                m_header->writeLine(iter);
-            }
-            m_header->writeLine();
-        }
-    }
-
-    // TODO: There are a lot of function calls and section below that expect a protected section.
-
-    m_header->Unindent();
-    m_header->writeLine("protected:");
-    m_header->Indent();
-
-    GenHdrEvents();
-
-    if (m_form_node->HasValue(prop_protected_class_methods))
-    {
-        tt_string_vector class_list(m_form_node->as_string(prop_protected_class_methods), '"',
-                                    tt::TRIM::both);
-        if (class_list.size())
-        {
-            m_header->writeLine();
-            for (auto& iter: class_list)
-            {
-                m_header->writeLine(iter);
-            }
-            m_header->writeLine();
-        }
-    }
-
-    if (!m_form_node->as_bool(prop_use_derived_class) && m_form_node->as_bool(prop_private_members))
-    {
-        m_header->Unindent();
-        m_header->writeLine();
-        m_header->writeLine("private:");
-        m_header->Indent();
-    }
-
-    CollectValidatorVariables(m_form_node, code_lines);
-    if (code_lines.size() || m_map_protected.size())
-    {
-        m_header->writeLine();
-        m_header->writeLine("// Validator variables");
-        if (code_lines.size())
-        {
-            m_header->writeLine();
-            WriteSetLines(m_header, code_lines);
-        }
-    }
-
-    if (m_map_protected.size())
-    {
-        for (auto& member: m_map_protected)
-        {
-            code.clear();
-            BeginPlatformCode(code, member.first);
-            m_header->writeLine(code);
-            for (const auto& code_line: member.second)
-            {
-                m_header->writeLine(code_line);
-            }
-            m_header->writeLine("#endif  // limited to specific platforms");
-        }
-        m_map_protected.clear();
-    }
-
-    CollectMemberVariables(m_form_node, Permission::Protected, code_lines);
-    generator->AddProtectedHdrMembers(code_lines);
-
-    if (code_lines.size() || m_map_protected.size())
-    {
-        m_header->writeLine();
-        m_header->writeLine("// Class member variables");
-        if (code_lines.size())
-        {
-            m_header->writeLine();
-            WriteSetLines(m_header, code_lines);
-        }
-    }
-
-    for (auto& member: m_map_protected)
-    {
-        code.clear();
-        BeginPlatformCode(code, member.first);
-        m_header->writeLine(code);
-        for (const auto& code_line: member.second)
-        {
-            m_header->writeLine(code_line);
-        }
-        m_header->writeLine("#endif  // limited to specific platforms");
-    }
-
-    if (m_form_node->HasValue(prop_class_members))
-    {
-        tt_string_vector class_list(m_form_node->as_string(prop_class_members), '"',
-                                    tt::TRIM::both);
-        m_header->writeLine();
-        for (auto& iter: class_list)
-        {
-            m_header->writeLine(iter);
-        }
-    }
-
-    m_header->Unindent();
-
-    if (not m_form_node->as_bool(prop_no_closing_brace))
-    {
-        m_header->writeLine("};");
     }
 }
 
@@ -1594,25 +1114,6 @@ void CppCodeGenerator::GenerateCppHandlers()
     }
 }
 
-void CppCodeGenerator::GenCppValidatorFunctions(Node* node)
-{
-    ASSERT(m_language == GEN_LANG_CPLUSPLUS);
-
-    if (node->HasValue(prop_validator_variable))
-    {
-        auto result = GenGetSetCode(node);
-        if (result)
-        {
-            m_header->writeLine(result.value());
-        }
-    }
-
-    for (const auto& child: node->get_ChildNodePtrs())
-    {
-        GenCppValidatorFunctions(child.get());
-    }
-}
-
 void CppCodeGenerator::GenUnhandledEvents(EventVector& events)
 {
     ASSERT_MSG(events.size(), "GenUnhandledEvents() shouldn't be called if there are no events");
@@ -1807,117 +1308,6 @@ void CppCodeGenerator::GenUnhandledEvents(EventVector& events)
     }
 }
 
-// This is called in gen_base.cpp, so it cannot be part of the CppCodeGenerator class.
-void CppCodeGenerator::GenCppValVarsBase(const NodeDeclaration* declaration, Node* node,
-                                         std::set<std::string>& code_lines)
-{
-    ASSERT(m_language == GEN_LANG_CPLUSPLUS);
-
-    if (const auto& var_name = node->as_string(prop_validator_variable); var_name.size())
-    {
-        if (auto val_data_type = node->get_ValidatorDataType(); val_data_type.size())
-        {
-            tt_string code;
-            code << val_data_type << ' ' << var_name;
-
-            if (val_data_type == "bool")
-            {
-                auto prop = node->get_PropPtr(prop_checked);
-                bool bState = (prop && prop->as_bool());
-                if (!prop)
-                {
-                    prop = node->get_PropPtr(prop_initial_state);
-                    if (prop && prop->as_string() == "wxCHK_CHECKED")
-                    {
-                        bState = true;
-                    }
-                }
-                code << " { " << (bState ? "true" : "false") << " };";
-            }
-            else if (val_data_type.contains("int") || val_data_type.contains("short") ||
-                     val_data_type.contains("long") || val_data_type.contains("double") ||
-                     val_data_type.contains("float"))
-            {
-                auto* prop = node->get_PropPtr(prop_value);
-                if (!prop)
-                {
-                    prop = node->get_PropPtr(prop_initial);
-                }
-                if (!prop)
-                {
-                    prop = node->get_PropPtr(prop_selection);
-                }
-                if (!prop)
-                {
-                    prop = node->get_PropPtr(prop_position);
-                }
-                if (prop && prop->as_string().size())
-                {
-                    code << " { " << prop->as_string() << " };";
-                }
-                else
-                {
-                    code << ';';
-                }
-            }
-            else if (val_data_type == "wxString" || val_data_type == "wxFileName")
-            {
-                const auto& value = node->as_string(prop_value);
-                if (value.size())
-                {
-                    code << " { " << GenerateQuotedString(value) << " };";
-                }
-                else
-                {
-                    code << ';';
-                }
-            }
-            // BUGBUG: [Randalphwa - 07-31-2023] We need to handle wxArrayInt
-
-            // REVIEW: [Randalphwa - 08-30-2025] I *think* the bug is still valid since we do have a
-            // wxArrayInt val_data_type, however for 3.3.x code, it needs to be replaced with std::
-            // containers.
-            else
-            {
-                code << ';';
-            }
-
-            // Validator variables are always written to the protected: section even if the
-            // node variable is marked as public:
-
-            if (node->HasProp(prop_platforms) &&
-                node->as_string(prop_platforms) != "Windows|Unix|Mac")
-            {
-                if (!m_map_protected.contains(node->as_string(prop_platforms)))
-                {
-                    m_map_protected[node->as_string(prop_platforms)] = std::set<tt_string>();
-                }
-                m_map_protected[node->as_string(prop_platforms)].insert(code);
-            }
-            // If node_container is non-null, it means the current node is within a container
-            // that has a conditional.
-            else if (auto* node_container = node->get_PlatformContainer(); node_container)
-            {
-                if (!m_map_protected.contains(node_container->as_string(prop_platforms)))
-                {
-                    m_map_protected[node_container->as_string(prop_platforms)] =
-                        std::set<tt_string>();
-                }
-                m_map_protected[node_container->as_string(prop_platforms)].insert(code);
-            }
-            else
-            {
-                code_lines.insert(code);
-            }
-        }
-    }
-
-    for (size_t i = 0; i < declaration->GetBaseClassCount(false); i++)
-    {
-        GenCppValVarsBase(declaration->GetBaseClass(i, false), node, code_lines);
-    }
-}
-
 // This should only be called to generate C++ code.
 void CppCodeGenerator::GenCppEnumIds(Node* class_node)
 {
@@ -1990,186 +1380,6 @@ void CppCodeGenerator::GenCppEnumIds(Node* class_node)
         m_header->writeLine();
         m_header->writeLine("};");
         m_header->writeLine();
-    }
-}
-
-void CppCodeGenerator::GenHdrEvents()
-{
-    ASSERT(m_language == GEN_LANG_CPLUSPLUS);
-
-    if (m_events.size() || m_ctx_menu_events.size())
-    {
-        std::set<tt_string> code_lines;
-
-        for (auto& event: m_events)
-        {
-            auto event_code = EventHandlerDlg::GetCppValue(event->get_value());
-            // Ignore lambda's and functions in another class
-            if (event_code.contains("[") || event_code.contains("::"))
-            {
-                continue;
-            }
-
-            tt_string code;
-
-            // If the form has a wxContextMenuEvent node, then the handler for the form's
-            // wxEVT_CONTEXT_MENU is a method of the base class and is not virtual.
-
-            if (event->getNode()->is_Form() && event->get_name() == "wxEVT_CONTEXT_MENU")
-            {
-                bool has_handler = false;
-
-                for (const auto& child: event->getNode()->get_ChildNodePtrs())
-                {
-                    if (child->is_Gen(gen_wxContextMenuEvent))
-                    {
-                        has_handler = true;
-                        break;
-                    }
-                }
-
-                if (has_handler)
-                {
-                    code << "void " << event_code << "("
-                         << event->get_EventInfo()->get_event_class() << "& event);";
-                    code_lines.insert(code);
-                    continue;
-                }
-            }
-            if (m_form_node->as_bool(prop_use_derived_class))
-            {
-                if (!m_form_node->as_bool(GenEnum::prop_pure_virtual_functions))
-                {
-                    code << "virtual void " << event_code << "("
-                         << event->get_EventInfo()->get_event_class()
-                         << "& event) { event.Skip(); }";
-                }
-                else
-                {
-                    code << "virtual void " << event_code << "("
-                         << event->get_EventInfo()->get_event_class() << "& event) = 0;";
-                }
-            }
-            else
-            {
-                code << "void " << event_code << "(" << event->get_EventInfo()->get_event_class()
-                     << "& event);";
-            }
-            code_lines.insert(code);
-        }
-
-        // Unlike the above code, there shouldn't be any wxEVT_CONTEXT_MENU events since
-        // m_ctx_menu_events should only contain menu items events.
-
-        for (const auto& event: m_ctx_menu_events)
-        {
-            auto event_code = EventHandlerDlg::GetCppValue(event->get_value());
-            // Ignore lambda's and functions in another class
-            if (event_code.contains("[") || event_code.contains("::"))
-            {
-                continue;
-            }
-
-            tt_string code;
-
-            if (m_form_node->as_bool(prop_use_derived_class))
-            {
-                if (!m_form_node->as_bool(GenEnum::prop_pure_virtual_functions))
-                {
-                    code << "virtual void " << event_code << "("
-                         << event->get_EventInfo()->get_event_class()
-                         << "& event) { event.Skip(); }";
-                }
-                else
-                {
-                    code << "virtual void " << event_code << "("
-                         << event->get_EventInfo()->get_event_class() << "& event) = 0;";
-                }
-            }
-            else
-            {
-                code << "void " << event_code << "(" << event->get_EventInfo()->get_event_class()
-                     << "& event);";
-            }
-
-            code_lines.insert(code);
-        }
-
-        if (code_lines.size())
-        {
-            m_header->writeLine();
-            if (m_form_node->as_bool(prop_use_derived_class))
-            {
-                m_header->writeLine(
-                    "// Virtual event handlers -- override them in your derived class");
-                m_header->writeLine();
-            }
-            else
-            {
-                m_header->writeLine("// Event handlers");
-                m_header->writeLine();
-            }
-            for (const auto& iter: code_lines)
-            {
-                m_header->writeLine(iter.subview());
-            }
-        }
-    }
-
-    if (m_map_conditional_events.size())
-    {
-        auto sort_events_by_handler = [](NodeEvent* event_a, NodeEvent* event_b)
-        {
-            return (event_a->get_value() < event_b->get_value());
-        };
-
-        if (m_events.empty() && m_ctx_menu_events.empty())
-        {
-            m_header->writeLine();
-            if (m_form_node->as_bool(prop_use_derived_class))
-            {
-                m_header->writeLine(
-                    "// Virtual event handlers -- override them in your derived class");
-            }
-            else
-            {
-                m_header->writeLine("// Event handlers");
-            }
-        }
-
-        for (auto& iter: m_map_conditional_events)
-        {
-            auto& events = iter.second;
-            std::sort(events.begin(), events.end(), sort_events_by_handler);
-            Code code(nullptr, GEN_LANG_CPLUSPLUS);
-            BeginPlatformCode(code, iter.first);
-            code.Eol();
-            for (auto& event: events)
-            {
-                auto event_code = EventHandlerDlg::GetCppValue(event->get_value());
-                // Ignore lambda's and functions in another class
-                if (event_code.contains("[") || event_code.contains("::"))
-                {
-                    continue;
-                }
-
-                if (m_form_node->as_bool(prop_use_derived_class))
-                {
-                    code << "virtual void " << event_code << "("
-                         << event->get_EventInfo()->get_event_class()
-                         << "& event) { event.Skip(); }";
-                }
-                else
-                {
-                    code << "void " << event_code << "("
-                         << event->get_EventInfo()->get_event_class() << "& event);";
-                }
-                code.Eol();
-            }
-            code << "#endif  // limited to specific platforms";
-            code.Eol();
-            m_header->writeLine(code);
-        }
     }
 }
 
@@ -2348,205 +1558,6 @@ void CppCodeGenerator::GenerateDataForm()
         m_header->writeLine("}\n");
     }
     ProjectData.WriteImagePostHeader(m_header);
-}
-
-void CppCodeGenerator::CollectValidatorVariables(Node* node, std::set<std::string>& code_lines)
-{
-    GenCppValVarsBase(node->get_NodeDeclaration(), node, code_lines);
-
-    for (const auto& child: node->get_ChildNodePtrs())
-    {
-        CollectValidatorVariables(child.get(), code_lines);
-    }
-}
-
-void CppCodeGenerator::CollectMemberVariables(Node* node, Permission perm,
-                                              std::set<std::string>& code_lines)
-{
-    if (m_form_node->is_Type(type_DocViewApp))
-    {
-        code_lines.emplace("wxFrame* m_frame;");
-        code_lines.emplace("wxDocManager* m_docManager;");
-        code_lines.emplace("wxMenuBar* m_menuBar;");
-
-        // Don't generate member variables for the Doc/View App node, they are not needed.
-        return;
-    }
-    if (m_form_node->is_Gen(gen_ViewTextCtrl))
-    {
-        code_lines.emplace("wxTextCtrl* m_text;");
-        return;
-    }
-    if (auto* prop = node->get_PropPtr(prop_class_access); prop)
-    {
-        if (prop->as_string() != "none")
-        {
-            if ((perm == Permission::Public && prop->as_string() == "public:") ||
-                (perm == Permission::Protected && prop->as_string() == "protected:"))
-            {
-                auto code = GetDeclaration(node);
-
-                auto ChangeClass = [&](const std::string& generic_class)
-                {
-                    if (auto pos = code.find('*'); pos != std::string::npos)
-                    {
-                        code = generic_class + code.substr(pos);
-                    }
-                };
-
-                if (node->is_Gen(gen_wxAnimationCtrl))
-                {
-                    if ((node->HasValue(prop_animation) &&
-                         node->as_string(prop_animation).contains(".ani", tt::CASE::either)) ||
-                        node->as_string(prop_subclass).starts_with("wxGeneric"))
-                    {
-                        // The generic version is required to display .ANI files on wxGTK.
-                        ChangeClass("wxGenericAnimationCtrl");
-                    }
-                }
-                else if (node->is_Gen(gen_wxHyperlinkCtrl))
-                {
-                    if (!node->as_bool(prop_underlined) ||
-                        node->as_string(prop_subclass).starts_with("wxGeneric"))
-                    {
-                        // If the underlined property is false, we need to use the generic
-                        // version.
-                        ChangeClass("wxGenericHyperlinkCtrl");
-                    }
-                }
-                else if (node->is_Gen(gen_wxStaticBitmap))
-                {
-                    if (node->as_string(prop_scale_mode) != "None" ||
-                        node->as_string(prop_subclass).starts_with("wxGeneric"))
-                    {
-                        // If we are using a scale mode, we must use wxGenericStaticBitmap.
-                        ChangeClass("wxGenericStaticBitmap");
-                    }
-                }
-                else if (node->is_Gen(gen_wxStaticText))
-                {
-                    if (node->as_string(prop_subclass).starts_with("wxGeneric") ||
-                        (node->as_bool(prop_markup) && node->as_int(prop_wrap) <= 0))
-                    {
-                        // If we are using markup or the wrap is <= 0, we must use
-                        // wxGenericStaticText.
-                        ChangeClass("wxGenericStaticText");
-                    }
-                }
-                else if (node->is_Gen(gen_wxCalendarCtrl))
-                {
-                    if (node->as_string(prop_subclass).starts_with("wxGeneric"))
-                    {
-                        ChangeClass("wxGenericCalendarCtrl");
-                    }
-                }
-                else if (node->is_Gen(gen_wxTreeCtrl))
-                {
-                    if (node->as_string(prop_subclass).starts_with("wxGeneric"))
-                    {
-                        ChangeClass("wxGenericTreeCtrl");
-                    }
-                }
-
-                else if (node->is_Gen(gen_wxTimer))
-                {
-                    // Remove the pointer, wxTimer needs to be a class.
-                    code.Replace("*", "");
-                }
-
-                if (code.empty() && node->is_Gen(gen_auitool))
-                {
-                    code += "wxAuiToolBarItem* " + node->as_string(prop_var_name) + ';';
-                }
-
-                if (code.size())
-                {
-                    if (node->HasProp(prop_platforms) &&
-                        node->as_string(prop_platforms) != "Windows|Unix|Mac")
-                    {
-                        if (perm == Permission::Public)
-                        {
-                            if (!m_map_public_members.contains(node->as_string(prop_platforms)))
-                            {
-                                m_map_public_members[node->as_string(prop_platforms)] =
-                                    std::set<tt_string>();
-                            }
-                            m_map_public_members[node->as_string(prop_platforms)].insert(code);
-                        }
-                        else
-                        {
-                            if (!m_map_protected.contains(node->as_string(prop_platforms)))
-                            {
-                                m_map_protected[node->as_string(prop_platforms)] =
-                                    std::set<tt_string>();
-                            }
-                            m_map_protected[node->as_string(prop_platforms)].insert(code);
-                        }
-                    }
-                    // If node_container is non-null, it means the current node is within a
-                    // container that has a conditional.
-                    else if (auto* node_container = node->get_PlatformContainer(); node_container)
-                    {
-                        if (perm == Permission::Public)
-                        {
-                            if (!m_map_public_members.contains(
-                                    node_container->as_string(prop_platforms)))
-                            {
-                                m_map_public_members[node_container->as_string(prop_platforms)] =
-                                    std::set<tt_string>();
-                            }
-                            m_map_public_members[node_container->as_string(prop_platforms)].insert(
-                                code);
-                        }
-                        else
-                        {
-                            if (!m_map_protected.contains(
-                                    node_container->as_string(prop_platforms)))
-                            {
-                                m_map_protected[node_container->as_string(prop_platforms)] =
-                                    std::set<tt_string>();
-                            }
-                            m_map_protected[node_container->as_string(prop_platforms)].insert(code);
-                        }
-                    }
-                    else
-                    {
-                        code_lines.insert(code);
-                    }
-                }
-            }
-        }
-    }
-
-    if (perm == Permission::Protected)
-    {
-        // StaticCheckboxBoxSizer and StaticRadioBtnBoxSizer have internal variables
-        if (node->HasValue(prop_checkbox_var_name) || node->HasValue(prop_radiobtn_var_name))
-        {
-            auto code = GetDeclaration(node);
-            if (code.size())
-            {
-                if (node->HasProp(prop_platforms) &&
-                    node->as_string(prop_platforms) != "Windows|Unix|Mac")
-                {
-                    if (!m_map_protected.contains(node->as_string(prop_platforms)))
-                    {
-                        m_map_protected[node->as_string(prop_platforms)] = std::set<tt_string>();
-                    }
-                    m_map_protected[node->as_string(prop_platforms)].insert(code);
-                }
-                else
-                {
-                    code_lines.insert(code);
-                }
-            }
-        }
-    }
-
-    for (const auto& child: node->get_ChildNodePtrs())
-    {
-        CollectMemberVariables(child.get(), perm, code_lines);
-    }
 }
 
 void CppCodeGenerator::CollectIncludes(Node* form, std::set<std::string>& set_src,
@@ -2749,43 +1760,6 @@ void CppCodeGenerator::GatherGeneratorIncludes(Node* node, std::set<std::string>
     {
         GatherGeneratorIncludes(child.get(), set_src, set_hdr);
     }
-}
-
-void CppCodeGenerator::WritePropHdrCode(Node* node, GenEnum::PropName prop)
-{
-    tt_string convert(node->as_string(prop));
-    convert.Replace("@@", "\n", tt::REPLACE::all);
-    tt_string_vector lines(convert, '\n', tt::TRIM::right);
-    bool initial_bracket = false;
-    for (auto& code: lines)
-    {
-        if (code.contains("}") && !code.contains("{"))
-        {
-            m_header->Unindent();
-        }
-        else if (!initial_bracket && code.contains("["))
-        {
-            initial_bracket = true;
-            m_header->Indent();
-        }
-
-        if (code.is_sameas("public:") || code.is_sameas("protected:") || code.is_sameas("private:"))
-        {
-            m_header->Unindent();
-            m_header->writeLine(code, indent::auto_no_whitespace);
-            m_header->Indent();
-        }
-        else
-        {
-            m_header->writeLine(code, indent::auto_no_whitespace);
-        }
-
-        if (code.contains("{") && !code.contains("}"))
-        {
-            m_header->Indent();
-        }
-    }
-    m_header->writeLine();
 }
 
 void CppCodeGenerator::WriteImagePostHeader()
